move profile picture to own file
